Updating entity classes

+++ src/main/java/com/example/bloggingapp/post/entity/PostTagId.java
@@ -15,6 +15,6 @@ import java.io.Serializable;
 @NoArgsConstructor
 public class PostTagId implements Serializable {
 
-    private Integer postId;
-    private Integer tagId;
+    private Long postId;
+    private Long tagId;
 }
